requirements.txt
django
djangorestframework
drf-spectacular
django-cors-headers
python-dotenv
djangorestframework-simplejwt
dj-database-url
psycopg2-binary